VariableFactory: move into `types` module; expose name generation
In preparation for deleting the old type-inference code, move the
variable factory into its own module. Also directly expose the ability
to get new names, so that our new code can use them.
